Refactored full text routines.
